agv wheels

+++ esp32_arm/src/AGV/mover_driver/mover_base.h
@@ -1,8 +1,8 @@
 # pragma once
-
+#include <stdint.h>
 
 class MoverBase{
     public:
         // virtual void SetSingleSpeed(int speed_in_percent);
-        virtual void SetForwdingSpeed(float forwarding_speed_in_percent, float angle);
+        virtual void SetForwdingSpeed(uint8_t forwarding_speed_in_percent, float angle);
 };
