+++ esp32_arm/src/AGV/mover_driver/mover_dual_wheel.h
@@ -8,7 +8,7 @@ class MoverDualWheel: public MoverBase{
         MoverDualWheel();
         void LinkLeftDriver(SingleWheelDriverBase* leftwheelDriver){this->__leftDriver=leftwheelDriver;};
         void LinkRightDriver(SingleWheelDriverBase* rightWheelDriver){this->__rightDriver=rightWheelDriver;};
-        void SetForwdingSpeed(float forwarding_speed_in_percent, float angle) override;
+        void SetForwdingSpeed(uint8_t forwarding_speed_in_percent, float angle) override;
 
     protected:
         SingleWheelDriverBase* __leftDriver;
